"Stable" version of python ide
Currently the client can edit and run python code in the text box. I've also created a scratch-bridge module that can be accessed in the python env. This will be the basis for the clients interaction with the Scratch renderer. Some uneeded files in this commit as I have already began work to switch Pyodide to run inside a web worker. This will allow for interrupt support.